Change to float64

+++ data/Dataset.py
@@ -66,10 +66,8 @@ class Dataset:
                     else:
                         features = speech_featurizer.compute_speech_features(audio_file)
                     labels = text_featurizer.compute_label_features(transcript)
-                    input_length = tf.convert_to_tensor(
-                        features.get_shape().as_list()[0])
-                    label_length = tf.convert_to_tensor(
-                        labels.get_shape().as_list()[0])
+                    input_length = tf.convert_to_tensor(features.get_shape().as_list()[0], dtype=tf.int64)
+                    label_length = tf.convert_to_tensor(labels.get_shape().as_list()[0], dtype=tf.int64)
                     if input_length < label_length:
                         continue
 
@@ -87,10 +85,10 @@ class Dataset:
             _gen_data,
             output_types=(
                 {
-                    "features": tf.float32,
-                    "input_length": tf.int32,
-                    "labels": tf.int32,
-                    "label_length": tf.int32
+                    "features": tf.float64,
+                    "input_length": tf.int64,
+                    "labels": tf.int64,
+                    "label_length": tf.int64
                 },
                 tf.int32
             ),
@@ -130,7 +128,7 @@ class Dataset:
         def _gen_data():
             for audio_file in self.entries:
                 features = speech_featurizer.compute_speech_features(audio_file)
-                input_length = tf.convert_to_tensor(features.get_shape().as_list()[0])
+                input_length = tf.convert_to_tensor(features.get_shape().as_list()[0], dtype=tf.int64)
                 yield (
                     {
                         "features": features,
@@ -143,8 +141,8 @@ class Dataset:
             _gen_data,
             output_types=(
                 {
-                    "features": tf.float32,
-                    "input_length": tf.int32
+                    "features": tf.float64,
+                    "input_length": tf.int64
                 },
                 tf.int32
             ),
